Multi File Uploads Now Possible!!
Sacrificed the replacement of "High" and "Low" readings with "400" and "40" for the ability to upload multiple files at once and have them display simultaneously on the graph. The overall average reading reflects this change. The impact on accuracy is minimal, but it's still worth exploring possible ways to make the interpolation work.